Movement now determined by dx (and soon dy)
I think this will mean that we don't need to store the time anymore when
jumping. If we take each iteration of the loop as a frame, the
information needed to determine the characters next position is just the
velocity, which is provided when the user presses a key.

TIL: the `->` macro in clojure DOES take partial functions, but they
need to be in extra brackets:

```
> (defn foo [a b] (+a b))
> (-> 2 (partial 1)) ;; returns the function
> (-> 2 ((partial 1)));; returns 3
```